# docs/api.yaml
openapi: 3.0.3
info:
  title: Limites Administrativos de Galiza
  description: Galiza Admin Boundaries
  version: 1.0.0

servers:
  - url: 'https://admin-levels-gz-luobwhahnq-nw.a.run.app'

paths:
  /comarcas:
    get:
      summary: Todas as comarcas
      responses:
        '200':
          description: Comarcas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AdminDivision'

  /comarcas/{idComarca}:
    get:
      summary: Obter unha division administrativa por código INE
      parameters:
        - in: path
          name: idComarca
          schema:
            type: integer
          required: true
          description: Código de comarca
      responses:
        '200':
          description: Array con todas as comarcas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminDivision'

  /divisions/{ineCode}:
    get:
      summary: Obter unha division administrativa por código INE
      parameters:
        - in: path
          name: ineCode
          schema:
            type: integer
          required: true
          description: Código INE
      responses:
        '200':
          description: Administrative division
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminDivision'

  /divisions/{ineCode}/{subDivision}:
    get:
      summary: Obter subdivions dunha division administrativa
      parameters:
        - in: path
          name: ineCode
          schema:
            type: integer
          required: true
          description: Código INE
        - in: path
          name: subDivision
          schema:
            type: string
            enum: [comarcas, concellos, parroquias, poboacions]
          required: true
          description: Subdivision administrativa
      responses:
        '200':
          description: Listado de subdivisiones administrativas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AdminDivision'

  /provincias:
    get:
      summary: Todas as provincias
      responses:
        '200':
          description: Provincias
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AdminDivision'

components:
  schemas:
    AdminDivision:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string